+++ lib/qc/modules/qc_tester/home/resources/testing_count_response.dart
@@ -5,16 +5,28 @@ part 'testing_count_response.g.dart';
 
 @JsonSerializable()
 class TestingCountResponse extends BaseResponse {
-  @JsonKey(name: "c")
-  int? testingDeviceCount;
 
-  @JsonKey(name: "ud")
-  int? lastUpdatedDate;
+  @JsonKey(name: "dt")
+  TestingCountData? testingDeviceData;
 
-  TestingCountResponse(this.testingDeviceCount, this.lastUpdatedDate, super.cashifyAlert, super.trackUrl);
+  TestingCountResponse(this.testingDeviceData, super.cashifyAlert, super.trackUrl);
 
   static TestingCountResponse fromJson(Map<String, dynamic> json) => _$TestingCountResponseFromJson(json);
 
   @override
   Map<String, dynamic> toJson() => _$TestingCountResponseToJson(this);
+}
+
+@JsonSerializable()
+class TestingCountData {
+
+  @JsonKey(name: "count")
+  int? testingDeviceCount;
+
+  @JsonKey(name: "lastUpdate")
+  int? lastUpdatedDate;
+
+  static TestingCountData fromJson(Map<String, dynamic> json) => _$TestingCountDataFromJson(json);
+
+  Map<String, dynamic> toJson() => _$TestingCountDataToJson(this);
 }