+++ lib/qc/modules/qc_tester/lob_devices/widgets/lob_device_detail_widget.dart
@@ -3,9 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/dialogs/show_mismatch_imei_dialog.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/dialogs/show_update_imei_dialog.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/providers/lob_device_scanner_provider.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/resources/brand_list_response.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/resources/device_category_id_type.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/resources/device_detail_response.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/lob_devices/resources/reasons.dart';
+import 'package:flutter_trc/src/common/widgets/dropdown_view_widget.dart';
 import 'package:flutter_trc/src/common/widgets/imei_scanner.dart';
 import 'package:flutter_trc/src/libraries/analytics/analytics_controller.dart';
 import 'package:flutter_trc/src/libraries/analytics/events/device_verify_popup_event.dart';
@@ -17,6 +19,7 @@ import 'package:flutter_trc/src/libraries/shared_preferences/app_preferences.dar
 import 'package:flutter_trc/src/modules/login/resources/login_types.dart';
 import 'package:image_picker/image_picker.dart';
 
+import '../dialogs/select_brand_bottom_sheet.dart';
 import '../l10n.dart';
 
 class LobDeviceDetailWidget extends StatefulWidget {
@@ -34,7 +37,7 @@ class LobDeviceDetailWidget extends StatefulWidget {
 class _LobDeviceDetailWidgetState extends State<LobDeviceDetailWidget> {
   final List<DropDownItem> _categoryList = [];
   DropDownItem? _selectedCategory;
-  DropDownItem? _selectedBrand;
+  BrandListData? _selectedBrand;
   bool _isImeiVerified = false;
   late bool _isRunImeiValidatorFlow = false;
 
@@ -121,16 +124,21 @@ class _LobDeviceDetailWidgetState extends State<LobDeviceDetailWidget> {
                 Flexible(
                   flex: 4,
                   fit: FlexFit.tight,
-                  child: CshDropDown(
-                    key: ValueKey(_selectedCategory?.id),
-                    items: provider.brandList,
-                    hintText: l10n.selectBrand,
-                    selectedItem: _selectedBrand,
-                    onChanged: (DropDownItem? value) {
-                      AnalyticsController.logEvent(SelectBrandEvent(widget.scannedData, value?.id));
-                      setState(() {
-                        _selectedBrand = value;
-                      });
+                  child: DropdownViewWidget(
+                    value: _selectedBrand?.brandName ?? l10n.selectBrand,
+                    isDataSelected: _selectedBrand != null,
+                    onPressed: () {
+                      selectBrandBottomSheet(
+                        context,
+                        provider.brandList ?? [],
+                        onBrandSelect: (brand) {
+                          Navigator.pop(context);
+                          AnalyticsController.logEvent(SelectBrandEvent(widget.scannedData, brand.brandId.toString()));
+                          setState(() {
+                            _selectedBrand = brand;
+                          });
+                        },
+                      );
                     },
                   ),
                 ),
@@ -184,7 +192,7 @@ class _LobDeviceDetailWidgetState extends State<LobDeviceDetailWidget> {
                     //     AutoSearchButtonClickedEvent(widget.scannedData, _selectedCategory?.id));
                     AnalyticsController.logEvent(
                         ManualSearchButtonClickedEvent(widget.scannedData, _selectedCategory?.id));
-                    widget.onSearchClicked(int.parse(_selectedBrand!.id!), int.parse(_selectedCategory!.id!));
+                    widget.onSearchClicked(_selectedBrand!.brandId!, int.parse(_selectedCategory!.id!));
                   }
                 : null,
           ),